popquiz-main/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
transformers
spacy
